Place Railhead Code in Module
